feat: pricing page revamp
