+++ app/components/pricing/cant-find.tsx
@@ -32,7 +32,7 @@ export default function CantFind() {
             </div>
             <div className={styles.bottom__section}>
                 <p>We&apos;d be happy to adapt our pricing to fit your use case! </p>
-                <a href="mailto:team@supertokens.com"><button><span>Get in touch</span></button></a>
+                <a role="button" href="mailto:team@supertokens.com"><span>Get in touch</span></a>
             </div>
         </section>
     );
