+++ app/styles/pricing/core-features-toggle.module.css
@@ -160,7 +160,7 @@
     font-size: 20px;
     font-weight: 600;
 
-    padding: 14px;
+    padding: 20px 0px 14px;
     width: 100%;
 
     display: flex;
